AP_SerialManager: implement new UART option bits

+++ libraries/AP_SerialManager/AP_SerialManager.h
@@ -212,7 +212,7 @@ private:
     bool protocol_match(enum SerialProtocol protocol1, enum SerialProtocol protocol2) const;
 
     // setup any special options
-    void set_options(uint8_t i);
+    void set_options(uint16_t i);
 };
 
 namespace AP {
